Add a flag to enable datadog

+++ backend/terraform/modules/analyzer/modules/engine_cluster/templates/user_data/engine.sh
@@ -103,17 +103,19 @@ sudo --login --set-home --user=ec2-user <<<"$(aws ecr get-login --no-include-ema
 sudo --login --set-home --user=ec2-user <<<'/usr/local/bin/docker-compose -f docker-compose.aws.yml -p artemis up -d'
 
 # Datadog Configurations:
-## Replace default Datadog configurations
-rm -f /etc/datadog-agent/datadog.yaml
-aws s3 cp "s3://${s3_bucket}/datadog/datadog.yaml" "/etc/datadog-agent/datadog.yaml"
+if [[ ${datadog_enabled} = "true" ]]; then
+  ## Replace default Datadog configurations
+  rm -f /etc/datadog-agent/datadog.yaml
+  aws s3 cp "s3://${s3_bucket}/datadog/datadog.yaml" "/etc/datadog-agent/datadog.yaml"
 
-# Update API Key
-ddog_api_key=$(aws secretsmanager get-secret-value --secret-id "${application}"/datadog-api-key --region "${aws_region}" --query SecretString --output text)
-sed -i "s/UPDATE_API_KEY/$ddog_api_key/" /etc/datadog-agent/datadog.yaml
+  # Update API Key
+  ddog_api_key=$(aws secretsmanager get-secret-value --secret-id "${application}"/datadog-api-key --region "${aws_region}" --query SecretString --output text)
+  sed -i "s/UPDATE_API_KEY/$ddog_api_key/" /etc/datadog-agent/datadog.yaml
 
-# Add dd-agent to docker users
-usermod -a -G docker dd-agent
+  # Add dd-agent to docker users
+  usermod -a -G docker dd-agent
 
-# Enable Datadog-agent
-systemctl enable datadog-agent
-systemctl restart datadog-agent
+  # Enable Datadog-agent
+  systemctl enable datadog-agent
+  systemctl restart datadog-agent
+fi
